+++ CHANGELOG.md
@@ -11,6 +11,7 @@ master
 0.7.1
 -----
 
+* Enhance `rake ember:install` to fully reinstall if necessary.
 * Resolve `ember` executable with full path within `node_modules`, instead of
   depending on the presence of `node_modules/.bin`. [#395]
 * Introduce the idea of `App#mountable?` and `App#to_rack` for handling deploys