+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 master
 ------
 
+0.7.2
+-----
+
 * Enhance `rake ember:install` to fully reinstall if necessary. [#396]
 * `EmberCli::Deploy::File` serves assets with Rails' `static_cache_control`
   value. [#403]